refactor: move PanAndZoom into RD.UI.Pancake
